tests/wpt/web-platform-tests/2dcontext/line-styles/2d.line.miter.invalid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /2dcontext/tools/gentest.py. -->
<title>Canvas test: 2d.line.miter.invalid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.line.miter.invalid</h1>
<p class="desc">Setting miterLimit to invalid values is ignored</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Setting miterLimit to invalid values is ignored");
_addTest(function(canvas, ctx) {

ctx.miterLimit = 1.5;
_assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");

ctx.miterLimit = 1.5;
ctx.miterLimit = 0;
_assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");

ctx.miterLimit = 1.5;
ctx.miterLimit = -1;
_assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");

ctx.miterLimit = 1.5;
ctx.miterLimit = Infinity;
_assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");

ctx.miterLimit = 1.5;
ctx.miterLimit = -Infinity;
_assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");

ctx.miterLimit = 1.5;
ctx.miterLimit = NaN;
_assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");

ctx.miterLimit = 1.5;
ctx.miterLimit = 'string';
_assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");

ctx.miterLimit = 1.5;
ctx.miterLimit = true;
_assertSame(ctx.miterLimit, 1, "ctx.miterLimit", "1");

ctx.miterLimit = 1.5;
ctx.miterLimit = false;
_assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");


});
</script>
